requirements.txt
aws-cdk-lib==2.175.1
constructs>=10.0.0,<11.0.0
requests